Adding compatibility with VIM session loading
When there is a NERDTree window open when a VIM session is saved, then
when the session is reloaded in a fresh instance of VIM there are errors
reported from the `autocmd`s.

This change updates the "OnEnter" autocmd to detect when the window
exists, but the `t:` variables were not set.  And it adds a new creator
function that should be called in that situation:  ReloadFromCWD.

The function creates the NERDTree object without creating the window.