Augmented "file(s) could not be loaded" error msg
This change also outputs the names of the files that could not be
loaded.  It does this before printing the same existing message that
just says how many.

So when you see the error in the status area, :messages can be used to
see the names of the offending files.