Adding new option: 'NERDTreeFindWithResolve'
This value was added to handle the case on Windows where the current
file *actually* lives on a network share, but was opened via a symbolic
link from one of the normal "drives". In that situation, resolving the
file's path yields the full "\\server\share\path\file.txt" notation
which cannot be found.

Without this option NERDTreeFind fully resolves the path every time.
The option was added with a default of "1" so the behavior stays the
same unless the user addes this option to their .vimrc file.